reorganized tutorial code to make it clearer.
grouped code in subpackages pipeline.core, pipeline.impl,
pipeline.util and app.

factored out superclasses SourceProcessor and SinkProcessor from,
respectively, File/CharSequenceReader and File/CharSequenceWriter
which abstract the source and sink setup code from the
feeding/receiving of source and sink data. the former go in core, the
latter in impl.

added TraceProcessor which dumps its input to System.out and passes
it on to the next processor in line without transforming it.

renamed PipelineAppMain to PipelineAppMain1.

provided a variant version of PipelineAppMain which uses
TraceProcessors to display the input, intermediate data and output
instead of teeing the output to a file via a TeeProcessor.